LimberJack.jl

Core

Core performs the main computations of LimberJack.jl. When using LimberJack.jl, the first step is to create an instance of the Cosmology structure. This is as easy as calling:

    using LimberJack
    cosmology = Cosmology()

This will generate the an instance of Cosmology given the vanilla $\Lambda$CDM cosmology of CCL. Cosmology() then computes the value of the comoving distance, the growth factor, the growth rate and matter power spectrum at an array of values and generates interpolators for said quantites. The user can acces the value of these interpolator at an arbitrary input using the public functions of the model. Moreover, the Cosmology structure has its own type called Cosmology.

LimberJack.Settings — Type

Settings(cosmo_type, nz, nz_pk, nk, tk_mode, pk_mode, custom_Dz)

Cosmology constructor settings structure.

Arguments:

cosmo_type::Type : type of cosmological parameters.

nz::Int : number of nodes in the general redshift array.

nz_pk::Int : number of nodes in the redshift array used for matter power spectrum.

nk::Int: number of nodes in the matter power spectrum.

tk_mode::String : choice of transfer function.

Pk_mode::String : choice of matter power spectrum.

custom_Dz::Any : custom growth factor.

Returns:

Settings : cosmology settings.

LimberJack.CosmoPar — Type

CosmoPar(Ωm, Ωb, h, n_s, σ8, θCMB, Ωr, ΩΛ)

Cosmology parameters structure.

Arguments:

Ωm::Dual : cosmological matter density.

Ωb::Dual : cosmological baryonic density.

h::Dual : reduced Hubble parameter.

n_s::Dual : spectral index.

σ8::Dual: variance of the matter density field in a sphere of 8 Mpc.

θCMB::Dual : CMB temperature.

Ωr::Dual : cosmological radiation density.

ΩΛ::Dual : cosmological dark energy density.

Returns:

CosmoPar : cosmology parameters structure.

LimberJack.Cosmology — Type

Cosmology

Base cosmology structure.

Arguments:

settings::MutableStructure : cosmology constructure settings.

cpar::MutableStructure : cosmological parameters.

chi::Dual : comoving distance array.

z_of_chi::Dual : redshift of comoving distance array.

chi_max::Dual : upper bound of comoving distance array.

chi_LSS::Dual : comoving distance to suface of last scattering.

Dz::Dual : growth factor.

fs8z::Dual : fs8.

PkLz0::Dual : interpolator of log primordial power spectrum over k-scales.

Pk::Dual : matter power spectrum.

Returns:

CosmoPar : cosmology parameters structure.

LimberJack.Ez — Function

Ez(cosmo::Cosmology, z)

Given a Cosmology instance, it returns the expansion rate (H(z)/H0).

Arguments:

cosmo::Cosmology : cosmology structure

z::Dual : redshift

Returns:

Ez::Dual : expansion rate

LimberJack.Hmpc — Function

Hmpc(cosmo::Cosmology, z)

Given a Cosmology instance, it returns the expansion history (H(z)) in Mpc.

Arguments:

cosmo::Cosmology : cosmology structure

z::Dual : redshift

Returns:

Hmpc::Dual : expansion rate

LimberJack.comoving_radial_distance — Function

comoving_radial_distance(cosmo::Cosmology, z)

Given a Cosmology instance, it returns the comoving radial distance.

Arguments:

cosmo::Cosmology : cosmology structure

z::Dual : redshift

Returns:

Chi::Dual : comoving radial distance

LimberJack.growth_factor — Function

growth_factor(cosmo::Cosmology, z)

Given a Cosmology instance, it returns the growth factor (D(z) = log(δ)).

Arguments:

cosmo::Cosmology : cosmological structure

z::Dual : redshift

Returns:

Dz::Dual : comoving radial distance

LimberJack.growth_rate — Function

growth_rate(cosmo::Cosmology, z)

Given a Cosmology instance, it returns growth rate.

Arguments:

cosmo::Cosmology : cosmology structure

z::Dual : redshift

Returns:

f::Dual : f

LimberJack.sigma8 — Function

sigma8(cosmo::Cosmology, z)

Given a Cosmology instance, it returns s8.

Arguments:

cosmo::Cosmology : cosmological structure

z::Dual : redshift

Returns:

s8::Dual : comoving radial distance

LimberJack.fs8 — Function

fs8(cosmo::Cosmology, z)

Given a Cosmology instance, it returns fs8.

Arguments:

cosmo::Cosmology : cosmology structure

z::Dual : redshift

Returns:

fs8::Dual : fs8

LimberJack.lin_Pk — Function

lin_Pk(cosmo::Cosmology, k, z)

Given a Cosmology instance, it returns the linear matter power spectrum (P(k,z))

Arguments:

cosmo::Cosmology : cosmology structure

k::Dual : scale

z::Dual : redshift

Returns:

Pk::Dual : linear matter power spectrum

LimberJack.nonlin_Pk — Function

nonlin_Pk(cosmo::Cosmology, k, z)

Given a Cosmology instance, it returns the non-linear matter power spectrum (P(k,z)) using the Halofit fitting formula (arXiv:astro-ph/0207664).

Arguments:

cosmo::Cosmology : cosmology structure

k::Dual : scale

z::Dual : redshift

Returns:

Pk::Dual : non-linear matter power spectrum

Tracers

In LimberJack.jl each tracer is structure containing at least of three fields:

wint: an interpolator between chis and warr.

F: an interpolator between chis and warr.

On top of these, tracers might contain the value of any nuisance parameter associated with them. Moreover, tracers within LimberJack.jl tracer objects have their own type called Tracer.

LimberJack.NumberCountsTracer — Type

NumberCountsTracer(warr, chis, wint, b, lpre)

Number counts tracer structure. Arguments:

wint::Interpolation : interpolation of the radial kernel over comoving distance.

F::Function : prefactor.

Returns:

NumberCountsTracer::NumberCountsTracer : Number counts tracer structure.

LimberJack.WeakLensingTracer — Type

WeakLensingTracer(warr, chis, wint, b, lpre)

Weak lensing tracer structure. Arguments:

wint::Interpolation : interpolation of the radial kernel over comoving distance.

F::Function : prefactor.

Returns:

WeakLensingTracer::WeakLensingTracer : Weak lensing tracer structure.

LimberJack.CMBLensingTracer — Type

CMBLensingTracer(warr, chis, wint, lpre)

CMB lensing tracer structure. Arguments:

wint::Interpolation : interpolation of the radial kernel over comoving distance.

F::Function : prefactor.

Returns:

CMBLensingTracer::CMBLensingTracer : CMB lensing tracer structure.

Spectra

Performs the computation of the angular power spectra of any two tracers.

LimberJack.Cℓintegrand — Function

Cℓintegrand(cosmo::Cosmology, t1::Tracer, t2::Tracer, logk, ℓ)

Returns the integrand of the angular power spectrum. Arguments:

cosmo::Cosmology : cosmology structure.

t1::Tracer : tracer structure.

t2::Tracer : tracer structure.

logk::Vector{Float} : log scale array.

ℓ::Float : multipole.

Returns:

integrand::Vector{Real} : integrand of the angular power spectrum.

LimberJack.angularCℓs — Function

angularCℓs(cosmo::Cosmology, t1::Tracer, t2::Tracer, ℓs)

Returns the angular power spectrum. Arguments:

cosmo::Cosmology : cosmology structure.

t1::Tracer : tracer structure.

t2::Tracer : tracer structure.

ℓs::Vector{Float} : multipole array.

Returns:

Cℓs::Vector{Real} : angular power spectrum.

Data Utils

Parses sacc and YAML files.

Missing docstring.

Missing docstring for LimberJack.make_data. Check Documenter's build log for details.

Theory

Performs the computation of complex theory vectors given sacc and YAML files.

Missing docstring.

Missing docstring for LimberJack.Theory. Check Documenter's build log for details.