LimberJack.jl

Core

LimberJack.Settings — Type

Settings(cosmo_type, nz, nz_pk, nk, tk_mode, pk_mode, custom_Dz)

Cosmology constructor settings structure.

Arguments:

cosmo_type::Type : type of cosmological parameters.

nz::Int : number of nodes in the general redshift array.

nz_pk::Int : number of nodes in the redshift array used for matter power spectrum.

nk::Int: number of nodes in the matter power spectrum.

tk_mode::String : choice of transfer function.

Pk_mode::String : choice of matter power spectrum.

custom_Dz::Any : custom growth factor.

Returns:

Settings : cosmology settings.

LimberJack.CosmoPar — Type

CosmoPar(Ωm, Ωb, h, n_s, σ8, θCMB, Ωr, ΩΛ)

Cosmology parameters structure.

Arguments:

Ωm::Dual : cosmological matter density.

Ωb::Dual : cosmological baryonic density.

h::Dual : reduced Hubble parameter.

n_s::Dual : spectral index.

σ8::Dual: variance of the matter density field in a sphere of 8 Mpc.

θCMB::Dual : CMB temperature.

Ωr::Dual : cosmological radiation density.

ΩΛ::Dual : cosmological dark energy density.

Returns:

CosmoPar : cosmology parameters structure.

LimberJack.Cosmology — Type

Cosmology(Settings, CosmoPar,
          ks, pk0, logk, dlogk,
          zs, chi, z_of_chi, chi_max, chi_LSS, Dz, PkLz0, Pk)

Base cosmology structure.

Arguments:

Settings::MutableStructure : cosmology constructure settings.

CosmoPar::Structure : cosmological parameters.

ks::Dual : scales array.

pk0::Dual: primordial matter power spectrum.

logk::Dual : log scales array.

dlogk::Dual : increment in log scales.

zs::Dual : redshift array.

chi::Dual : comoving distance array.

z_of_chi::Dual : redshift of comoving distance array.

chi_max::Dual : upper bound of comoving distance array.

chi_LSS::Dual : comoving distance to suface of last scattering.

Dz::Dual : growth factor.

PkLz0::Dual : interpolator of log primordial power spectrum over k-scales.

Pk::Dual : matter power spectrum.

Returns:

CosmoPar : cosmology parameters structure.

LimberJack.Ez — Function

Ez(cosmo::Cosmology, z)

Given a Cosmology instance, it returns the expansion rate (H(z)/H0).

Arguments:

cosmo::Cosmology : cosmology structure

z::Dual : redshift

Returns:

Ez::Dual : expansion rate

LimberJack.Hmpc — Function

Hmpc(cosmo::Cosmology, z)

Given a Cosmology instance, it returns the expansion history (H(z)) in Mpc.

Arguments:

cosmo::Cosmology : cosmology structure

z::Dual : redshift

Returns:

Hmpc::Dual : expansion rate

LimberJack.comoving_radial_distance — Function

comoving_radial_distance(cosmo::Cosmology, z)

Given a Cosmology instance, it returns the comoving radial distance.

Arguments:

cosmo::Cosmology : cosmology structure

z::Dual : redshift

Returns:

Chi::Dual : comoving radial distance

LimberJack.growth_factor — Function

growth_factor(cosmo::Cosmology, z)

Given a Cosmology instance, it returns the growth factor (D(z) = log(δ)).

Arguments:

cosmo::Cosmology : cosmological structure

z::Dual : redshift

Returns:

Dz::Dual : comoving radial distance

LimberJack.lin_Pk — Function

lin_Pk(cosmo::Cosmology, k, z)

Given a Cosmology instance, it returns the linear matter power spectrum (P(k,z))

Arguments:

cosmo::Cosmology : cosmology structure

k::Dual : scale

z::Dual : redshift

Returns:

Pk::Dual : linear matter power spectrum

LimberJack.nonlin_Pk — Function

nonlin_Pk(cosmo::Cosmology, k, z)

Given a Cosmology instance, it returns the non-linear matter power spectrum (P(k,z)) using the Halofit fitting formula (arXiv:astro-ph/0207664).

Arguments:

cosmo::Cosmology : cosmology structure

k::Dual : scale

z::Dual : redshift

Returns:

Pk::Dual : non-linear matter power spectrum

Emulator

LimberJack.Emulator — Type

Emulator()

Emulator structure.

Arguments:

alphas::Matrix : emulator's linear model for each k-scale

hypers::Matrix : hyperparameters for each k-scale

trans_cosmos::Vector{Array} : transformed cosmological parameters

training_karr::Vector{Float} : array of k-scales at which the emulator was trained

inv_chol_cosmos::Matrix : inverse Cholesky decomposition of training matter power spectra

mean_cosmos::Vector{Float} : mean training cosmology

mean_log_Pks::Vector{Float} : mean of the log training matter power spectra

std_log_Pks::Vector{Float} : standard deviation of the log training matter power spectrum

Returns:

emulator::Structure : emulator structure.

LimberJack.get_emulated_log_pk0 — Function

get_emulated_log_pk0(cosmo::CosmoPar)

Given a CosmoPar instance, emulates the value of the primordial matter power spectrum at the training k-scales.

Arguments:

CosmoPar::Stucture : cosmological paramters structure

Returns:

karr::Vector{Float} : array of training k-scales

log_pk0s::Vector{Dual} : array of emulated log matter power spectrum

Halofit

LimberJack.get_PKnonlin — Function

get_PKnonlin(cosmo::CosmoPar, z, k, PkLz0, Dzs, cosmo_type::DataType)

Uses the Halofit (arXiv:astro-ph/0207664) model to calculate the non-linear matter power spectrum.

Arguments:

CosmoPar::Stucture : cosmological paramters structure

z::Vector{Dual} : array of redshifts

k::Vector{Dual} : array of k-scales

PkLz0::Interpolator : log primordial power spectrum interpolator as a function of log k-scale

Dzs::Vector{Dual} : growth factor evaluated at redshift array

cosmo_type::Type : type of cosmological parameters

Returns:

pk_NL(logk, z)::Interpolator : non-linear matter spectrum as a function of log k-scale and redshift

Spectra

LimberJack.angularCℓs — Function

angularCℓs(cosmo::Cosmology, t1::Tracer, t2::Tracer, ℓs)

Returns the angular power spectrum.

Arguments:

cosmo::Cosmology : cosmology structure.

t1::Tracer : tracer structure.

t2::Tracer : tracer structure.

ℓs::Vector{Float} : multipole array.

Returns:

Cℓs::Vector{Real} : angular power spectrum.

Missing docstring.

Missing docstring for LimberJack.angularCℓ. Check Documenter's build log for details.

Tracers

LimberJack.NumberCountsTracer — Type

NumberCountsTracer(warr, chis, wint, b, lpre)

Number counts tracer structure.

Arguments:

warr::Vector{Dual} : radial kernel array.

chis::Vector{Dual} : comoving distances array of the radial kernel.

wint::Interpolation : interpolation of the radial kernel over comoving distance.

b::Dual: matter-galaxy bias.

lpre::Int : prefactor mark.

Returns:

NumberCountsTracer::NumberCountsTracer : Number counts tracer structure.

LimberJack.WeakLensingTracer — Type

WeakLensingTracer(warr, chis, wint, b, lpre)

Weak lensing tracer structure. Arguments:

warr::Vector{Dual} : radial kernel array.

chis::Vector{Dual} : comoving distances array of the radial kernel.

wint::Interpolation : interpolation of the radial kernel over comoving distance.

b::Dual: matter-galaxy bias.

Returns:

WeakLensingTracer::WeakLensingTracer : Weak lensing tracer structure.

LimberJack.CMBLensingTracer — Type

CMBLensingTracer(warr, chis, wint, lpre)

CMB lensing tracer structure.

Arguments:

warr::Vector{Dual} : radial kernel array.

chis::Vector{Dual} : comoving distances array of the radial kernel.

wint::Interpolation : interpolation of the radial kernel over comoving distance.

lpre::Int : prefactor mark.

Returns:

CMBLensingTracer::CMBLensingTracer : CMB lensing tracer structure.